+++ bin/nanocld/src/utils/resource.rs
@@ -62,10 +62,10 @@ async fn hook_delete_resource(
   resource: &Resource,
   pool: &Pool,
 ) -> HttpResult<()> {
-  let kind: ResourceKind =
-    SpecDb::get_version(&resource.kind, &resource.spec.version, pool)
-      .await?
-      .try_into()?;
+  let (kind, version) = ResourceDb::parse_kind(&resource.kind, pool).await?;
+  let kind: ResourceKind = SpecDb::get_version(&kind, &version, pool)
+    .await?
+    .try_into()?;
   log::debug!("hook_delete_resource kind: {kind:?}");
   if let Some(url) = &kind.data.url {
     let ctrl_client = CtrlClient::new(&kind.name, url);