+++ bin/ncproxy/src/cli.rs
@@ -3,8 +3,11 @@ use clap::Parser;
 #[derive(Parser)]
 pub struct Cli {
   /// Path to nginx config directory
+  #[clap(long, default_value = "/etc/nginx")]
+  pub nginx_dir: String,
+  /// Path to state directory
   #[clap(long)]
-  pub conf_dir: Option<String>,
+  pub state_dir: String,
 }
 
 #[cfg(test)]
@@ -13,10 +16,18 @@ mod tests {
 
   #[test]
   fn parse() {
-    let args = Cli::parse_from(["nanocl-ncproxy", "--conf-dir", "/etc/nginx"]);
-    assert_eq!(args.conf_dir, Some("/etc/nginx".into()));
-    let args = Cli::parse_from(["nanocl-ncproxy"]);
-    assert_eq!(args.conf_dir, None);
+    let args = Cli::parse_from([
+      "ncproxy",
+      "--nginx-dir",
+      "/test/nginx",
+      "--state-dir",
+      "/test/state",
+    ]);
+    assert_eq!(args.nginx_dir, "/test/nginx");
+    assert_eq!(args.state_dir, "/test/state");
+    let args = Cli::parse_from(["ncproxy", "--state-dir", "/test/state"]);
+    assert_eq!(args.nginx_dir, "/etc/nginx");
+    assert_eq!(args.state_dir, "/test/state");
     let _ = Cli::try_parse();
   }
 }