+++ bin/ncproxy/src/models/mod.rs
@@ -1,0 +1,7 @@
+mod store;
+mod system;
+mod template;
+
+pub use store::*;
+pub use system::*;
+pub use template::*;