+++ bin/ncproxy/changelog.md
@@ -5,6 +5,15 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## untagged
+
+### Changed
+
+- Added state dir
+- Reload task that can be canceled
+- Test file generation before creating the resource
+- Self manage nginx process
+- Metrics saving
 
 ## [0.9.0] - 2023-12-22